Moved transaction scope manager from agent to TransactionFrame type (to not collide with the .NET TransactionScope type). Type added to dependency injection, tests are updated and documentation.
