+++ src/Orleans.Transactions.TestKit.Base/TestRunners/FramedTransactionsTestRunner.cs
@@ -5,14 +5,14 @@ using FluentAssertions;
 
 namespace Orleans.Transactions.TestKit
 {
-    public abstract class ScopedTransactionsTestRunner : TransactionTestRunnerBase
+    public abstract class FramedTransactionsTestRunner : TransactionTestRunnerBase
     {
-        private readonly ITransactionAgent _transactionAgent;
+        private readonly ITransactionFrame _transactionFrame;
 
-        protected ScopedTransactionsTestRunner(IGrainFactory grainFactory, ITransactionAgent transactionAgent, Action<string> output)
+        protected FramedTransactionsTestRunner(IGrainFactory grainFactory, ITransactionFrame transactionFrame, Action<string> output)
             : base(grainFactory, output)
         {
-            _transactionAgent = transactionAgent;
+            _transactionFrame = transactionFrame;
         }
 
         public virtual async Task CreateTransactionScopeAndSetValue(string grainStates)
@@ -23,7 +23,7 @@ namespace Orleans.Transactions.TestKit
             // Act
             Func<Task> act = () => grain.Set(57);
 
-            await _transactionAgent.Transaction(TransactionOption.Create, async () =>
+            await _transactionFrame.RunScope(TransactionOption.Create, async () =>
                 // Assert
                 await act.Should().NotThrowAsync(because: "No failure expected"));
         }
@@ -36,7 +36,7 @@ namespace Orleans.Transactions.TestKit
             // Act
             Func<Task> act = () => grain.SetAndThrow(57);
 
-            await _transactionAgent.Transaction(TransactionOption.Create, async () =>
+            await _transactionFrame.RunScope(TransactionOption.Create, async () =>
                 // Assert
                 await act.Should().ThrowAsync<OrleansTransactionAbortedException>(because: "Failure expected"));
         }
@@ -49,7 +49,7 @@ namespace Orleans.Transactions.TestKit
             var grain = RandomTestGrain(grainStates);
 
             // Act
-            await _transactionAgent.Transaction(TransactionOption.Create, async () =>
+            await _transactionFrame.RunScope(TransactionOption.Create, async () =>
             {
                 await grain.Set(57);
                 result = await grain.Get();
@@ -67,13 +67,13 @@ namespace Orleans.Transactions.TestKit
             var grain = RandomTestGrain(grainStates);
 
             // Act
-            await _transactionAgent.Transaction(TransactionOption.Create, async () =>
+            await _transactionFrame.RunScope(TransactionOption.Create, async () =>
             {
                 await grain.Set(57);
 
                 try
                 {
-                    await _transactionAgent.Transaction(TransactionOption.Create, async () => await grain.SetAndThrow(67));
+                    await _transactionFrame.RunScope(TransactionOption.Create, async () => await grain.SetAndThrow(67));
                 }
                 catch
                 { }
