+++ src/Orleans.Transactions/DisabledTransactionAgent.cs
@@ -8,5 +8,4 @@ internal class DisabledTransactionAgent : ITransactionAgent
     public Task Abort(TransactionInfo transactionInfo) => throw new OrleansTransactionsDisabledException();
     public Task<(TransactionalStatus Status, Exception exception)> Resolve(TransactionInfo transactionInfo) => throw new OrleansTransactionsDisabledException();
     public Task<TransactionInfo> StartTransaction(bool readOnly, TimeSpan timeout) => throw new OrleansStartTransactionFailedException(new OrleansTransactionsDisabledException());
-    public Task Transaction(TransactionOption transactionOption, Func<Task> scope) => throw new OrleansTransactionsDisabledException();
 }
